Fix deploy from VS issue on iOS app
Issue was caused by a linked image file that didn't get bundled together
when deployed form VS to iOS build host. Simply adding the image as a
bundledresource and not linking it fixes it for now. Bug:
https://bugzilla.xamarin.com/show_bug.cgi?id=18843

Additionally ensure that iPhoneSimulator will deploy a if selected. See
changes in XamarinStore.sln